requirements.txt
# RVC Inference API 依赖

# 核心框架
torch>=2.0.0
torchaudio>=2.0.0

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# 音频处理
librosa>=0.9.0
soundfile>=0.12.0
scipy>=1.9.0
pyworld>=0.3.0
praat-parselmouth>=0.4.0
pydub>=0.25.0
ffmpeg-python>=0.2.0
resampy>=0.4.0
audioread

# 深度学习辅助
numpy>=1.23.0
numba>=0.56.0
faiss-cpu>=1.7.0
fairseq>=0.12.0
Cython

# 音高提取
torchcrepe>=0.0.20
torchfcpe>=0.0.4

# 其他工具
python-dotenv>=1.0.0
tqdm>=4.65.0
scikit-learn>=1.2.0
PyYAML>=6.0
Pillow>=9.1.0
joblib>=1.1.0
llvmlite>=0.39.0

# 可选ONNX支持
onnxruntime; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'
